requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
discord.py==1.3.1
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.8
multidict==4.7.4
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.5
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
websockets==8.1
yarl==1.4.2